Refactor fullScreenWarning.js with Aphrodite
Closes #7250

Auditors:

Test Plan:
1. Open any youtube video
2. Make the player fullscreen
