+++ js/components/frame.js
@@ -16,7 +16,7 @@ const UrlUtil = require('../lib/urlutil')
 const messages = require('../constants/messages')
 const contextMenus = require('../contextMenus')
 const ipc = require('electron').ipcRenderer
-const FullScreenWarning = require('./fullScreenWarning')
+const FullScreenWarning = require('../../app/renderer/components/fullScreenWarning')
 const debounce = require('../lib/debounce')
 const getSetting = require('../settings').getSetting
 const config = require('../constants/config')
