+++ less/window.less
@@ -66,19 +66,6 @@ html,
   top: 0;
   left: 0;
 
-  .fullScreenModeWarning {
-    background: #3a3a3a;
-    color: white;
-    font-size: larger;
-    left: 0;
-    margin-left: auto;
-    margin-right: auto;
-    padding: 20px;
-    position: absolute;
-    right: 0;
-    text-align: center;
-  }
-
   // Needed to be able to click and drag to select content in pages.
   -webkit-user-select: none;
